requirements.txt
tensorflow
tensorflow_hub
tensorflow_text
tensorflow-datasets
tf-models-official
tensorboard
pandas
numpy
matplotlib
coinbase
pytrends
tweepy
tqdm
kaggle